Package improvements, removing AdvancedTiledView
