+++ src/main/java/net/imagej/modelzoo/consumer/model/Model.java
@@ -27,83 +27,29 @@
  * #L%
  */
 
-package net.imagej.modelzoo.consumer.network.model;
+package net.imagej.modelzoo.consumer.model;
 
-import net.imagej.axis.AxisType;
-import net.imagej.modelzoo.consumer.tiling.Tiling;
+import org.scijava.Disposable;
 
-public class ModelZooAxis {
+import java.io.FileNotFoundException;
+import java.util.List;
 
-	private Integer min;
-	private Integer halo;
-	private Double scale;
-	private Integer offset;
-	private Tiling.TilingAction tiling;
-	private Long actual;
-	private AxisType type;
-	private Integer step;
+public interface Model extends Disposable {
 
-	public Integer getMin() {
-		return min;
+	boolean loadModel(String pathOrURL, String modelName)
+			throws FileNotFoundException;
+
+	List<InputImageNode<?>> getInputNodes();
+
+	List<OutputImageNode<?, ?>> getOutputNodes();
+
+	boolean isInitialized();
+
+	default void loadLibrary() {
 	}
 
-	public Integer getHalo() {
-		return halo;
-	}
+	boolean libraryLoaded();
 
-	public Double getScale() {
-		return scale;
-	}
+	void predict();
 
-	public Integer getOffset() {
-		return offset;
-	}
-
-	public Tiling.TilingAction getTiling() {
-		return tiling;
-	}
-
-	public Long getActual() {
-		return actual;
-	}
-
-	public ModelZooAxis(AxisType axisType) {
-		type = axisType;
-	}
-
-	public AxisType getType() {
-		return type;
-	}
-
-	public void setMin(Integer min) {
-		this.min = min;
-	}
-
-	public void setHalo(Integer halo) {
-		this.halo = halo;
-	}
-
-	public void setScale(Double scale) {
-		this.scale = scale;
-	}
-
-	public void setOffset(Integer offset) {
-		this.offset = offset;
-	}
-
-	public void setTiling(Tiling.TilingAction tiling) {
-		this.tiling = tiling;
-	}
-
-	public void setActual(Long actual) {
-		this.actual = actual;
-	}
-
-	public Integer getStep() {
-		return step;
-	}
-
-	public void setStep(Integer step) {
-		this.step = step;
-	}
 }
