+++ src/main/java/net/imagej/modelzoo/consumer/model/ModelZooNode.java
@@ -1,4 +1,4 @@
-package net.imagej.modelzoo.consumer.network.model;
+package net.imagej.modelzoo.consumer.model;
 
 public interface ModelZooNode<T> {
 	void setName(String name);
