--- src/main/java/net/imagej/modelzoo/consumer/tiling/AdvancedTiledView.java
@@ -1,64 +1,0 @@
-/*-
- * #%L
- * ImageJ ModelZoo Consumer
- * %%
- * Copyright (C) 2019 MPI-CBG
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-
-package net.imagej.modelzoo.consumer.tiling;
-
-import net.imagej.axis.AxisType;
-import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.TiledView;
-import net.imglib2.type.numeric.RealType;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-class AdvancedTiledView<TI extends RealType<TI>, TO extends RealType<TO>> extends TiledView<TI> {
-
-	private final Map<AxisType, Long> originalDims;
-	private final List<RandomAccessibleInterval<TO>> processedTiles;
-	// protected int blockMultiple;
-	// protected long blockWidth;
-
-	AdvancedTiledView(final RandomAccessibleInterval<TI> source,
-	                  final long[] blockSize, final long[] overlap) {
-		super(source, blockSize, overlap);
-		processedTiles = new ArrayList<>();
-		originalDims = new HashMap<>();
-	}
-
-	Map<AxisType, Long> getOriginalDims() {
-		return originalDims;
-	}
-
-	List<RandomAccessibleInterval<TO>> getProcessedTiles() {
-		return processedTiles;
-	}
-
-}
